fix: pull progress logging
